requirements.txt
google-cloud-aiplatform >= 1.31.0
flask
google-cloud-logging
python-dotenv